fixed time stamp

+++ Assets/Scripts/TriggerSoundTarget.cs
@@ -41,7 +41,7 @@ public class TriggerSoundTarget : MonoBehaviour
                 AudioSource.PlayClipAtPoint(soundTrigger, transform.position, 1.0f);
                 sound_has_played = true;
                 System.DateTime epochStart = new System.DateTime(1970, 1, 1, 0, 0, 0, System.DateTimeKind.Utc);
-                float cur_time = (float)(System.DateTime.UtcNow - epochStart).TotalSeconds;
+                int cur_time = (int)(System.DateTime.UtcNow - epochStart).TotalSeconds;
                 Debug.Log("Triggered Tar:" + colliding_object.transform.position.ToString("F4") + " at Epoch Time " + cur_time.ToString());
             }
         }
